refactor(api): move protocol implementation classes to protocol_api._core
Re: RCORE-79
