+++ api/src/opentrons/protocol_api/_core/__init__.py
@@ -1,0 +1,7 @@
+"""Core protocol logic interfaces and implementations.
+
+This module provide facades to different protocol execution cores,
+and is an internal implementation detail of the Python Protocol API.
+
+Nothing defined in `opentrons.protocol_api._core` may be considered public.
+"""
