+++ api/src/opentrons/protocol_api/_core/instrument.py
@@ -8,7 +8,8 @@ import typing
 from opentrons import types
 from opentrons.hardware_control.dev_types import PipetteDict
 from opentrons.protocols.api_support.util import Clearances, PlungerSpeeds, FlowRates
-from opentrons.protocols.context.well import WellImplementation
+
+from .well import WellImplementation
 
 
 class AbstractInstrument(ABC):
