+++ api/src/opentrons/protocol_api/_core/labware.py
@@ -6,10 +6,11 @@ from typing import List, Dict, Optional
 from opentrons.protocols.geometry.deck_item import DeckItem
 from opentrons.protocols.geometry.labware_geometry import AbstractLabwareGeometry
 from opentrons.protocols.api_support.tip_tracker import TipTracker
-from opentrons.protocols.context.well import WellImplementation
 from opentrons.protocols.api_support.well_grid import WellGrid
 from opentrons.types import Point
 from opentrons_shared_data.labware.dev_types import LabwareParameters, LabwareDefinition
+
+from .well import WellImplementation
 
 
 class AbstractLabware(DeckItem):
@@ -64,7 +65,7 @@ class AbstractLabware(DeckItem):
         ...
 
     @abstractmethod
-    def set_tip_length(self, length: float):
+    def set_tip_length(self, length: float) -> None:
         ...
 
     @abstractmethod
