+++ api/src/opentrons/protocols/api_support/util.py
@@ -22,7 +22,7 @@ from opentrons.protocols.api_support.types import APIVersion
 from opentrons.hardware_control.types import Axis
 
 if TYPE_CHECKING:
-    from opentrons.protocols.context.instrument import AbstractInstrument
+    from opentrons.protocol_api._core.instrument import AbstractInstrument
     from opentrons.protocol_api.labware import Well, Labware
     from opentrons.protocols.geometry.deck import Deck
