--- api/src/opentrons/protocols/context/__init__.py
@@ -1,4 +1,0 @@
-"""
-This package defines the interfaces that implement the ProtocolContext,
- InstrumentContext, and related classes.
-"""
