+++ api/tests/opentrons/protocol_api/test_labware.py
@@ -8,8 +8,8 @@ from decoy import Decoy
 from opentrons.protocol_api import labware, MAX_SUPPORTED_VERSION
 from opentrons.protocols.geometry import module_geometry
 from opentrons.protocols.geometry.well_geometry import WellGeometry
-from opentrons.protocols.context.protocol_api.labware import LabwareImplementation
-from opentrons.protocols.context.well import WellImplementation
+from opentrons.protocol_api._core.protocol_api.labware import LabwareImplementation
+from opentrons.protocol_api._core.well import WellImplementation
 from opentrons.protocols.labware.definition import _get_parent_identifier
 
 from opentrons_shared_data import load_shared_data
@@ -20,7 +20,7 @@ from opentrons.types import Point, Location
 from opentrons.hardware_control.modules.types import ModuleType, MagneticModuleModel
 from opentrons.protocols.api_support.types import APIVersion
 from opentrons.protocols.api_support.util import APIVersionError
-from opentrons.protocols.context.labware import AbstractLabware
+from opentrons.protocol_api._core.labware import AbstractLabware
 from opentrons.protocols.geometry.deck import Deck
 from opentrons.protocols.geometry.module_geometry import ModuleGeometry
