+++ api/tests/opentrons/protocols/api_support/test_tip_tracker.py
@@ -2,7 +2,7 @@ from typing import List
 
 import pytest
 from opentrons.protocols.api_support.tip_tracker import TipTracker
-from opentrons.protocols.context.well import WellImplementation
+from opentrons.protocol_api._core.well import WellImplementation
 from opentrons.protocols.api_support.well_grid import WellGrid
